refactor: extract chore-mutation primitives into actions.py
services.py had grown to 770 lines (over the project's 500-line max
target). The shared mutation primitives — async_complete_chore,
async_uncomplete_chore, async_apply_completed_cleared_at, and the
calendar event-listener notifier — move into actions.py. services.py
keeps the handler glue, schemas, registration, and resolve helpers.
